refactor: Move real BTC fixtures into fedimint-testing
