+++ integrationtests/Cargo.toml
@@ -15,12 +15,10 @@ anyhow = "1.0.66"
 assert_matches = "1.5.0"
 async-trait = "0.1.64"
 bitcoin = "0.29.2"
-bitcoincore-rpc = "0.16.0"
 cln-rpc = "0.1.1"
 erased-serde = "0.3"
 futures = "0.3.24"
 itertools = "0.10.5"
-lazy_static = "1.4.0"
 lightning-invoice = "0.21.0"
 ln-gateway = { path = "../gateway/ln-gateway" }
 lightning = "0.0.113"
